moved ALLOWDELETE types from entity_service to helpers where DeleteEntityByID is located
